# docs/openapi.json
{
  "components": {
    "schemas": {
      "ChangeCourseCapacityRequest": {
        "description": "Request body for the Change Course Capacity command.",
        "properties": {
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          }
        },
        "required": [
          "capacity"
        ],
        "title": "ChangeCourseCapacityRequest",
        "type": "object"
      },
      "CommandResponse": {
        "description": "Response body reporting the outcome of a successful command.",
        "properties": {
          "event_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Event Ids",
            "type": "array"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "required": [
          "event_ids",
          "position"
        ],
        "title": "CommandResponse",
        "type": "object"
      },
      "CourseCatalogueEntryResponse": {
        "description": "Response body for a single course in the catalogue.",
        "properties": {
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "number_of_subscriptions": {
            "title": "Number Of Subscriptions",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "title",
          "capacity",
          "number_of_subscriptions"
        ],
        "title": "CourseCatalogueEntryResponse",
        "type": "object"
      },
      "CourseCatalogueResponse": {
        "description": "Response body for the Course Catalogue query.",
        "properties": {
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseCatalogueEntryResponse"
            },
            "title": "Courses",
            "type": "array"
          }
        },
        "required": [
          "courses"
        ],
        "title": "CourseCatalogueResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Response body for an operational probe.",
        "properties": {
          "status": {
            "enum": [
              "alive",
              "ready"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "RegisterCourseRequest": {
        "description": "Request body for the Register Course command.",
        "properties": {
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "title",
          "capacity"
        ],
        "title": "RegisterCourseRequest",
        "type": "object"
      },
      "StudentCourseSubscriptionsResponse": {
        "description": "Response body for the Student Course Subscriptions query.",
        "properties": {
          "course_limit": {
            "title": "Course Limit",
            "type": "integer"
          },
          "courses": {
            "items": {
              "type": "string"
            },
            "title": "Courses",
            "type": "array"
          },
          "student_id": {
            "title": "Student Id",
            "type": "string"
          },
          "subscription_count": {
            "title": "Subscription Count",
            "type": "integer"
          }
        },
        "required": [
          "student_id",
          "subscription_count",
          "course_limit",
          "courses"
        ],
        "title": "StudentCourseSubscriptionsResponse",
        "type": "object"
      },
      "StudentRegisteredWebhook": {
        "description": "Payload of a Student Registered notification from the registrar.",
        "properties": {
          "course_limit": {
            "title": "Course Limit",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "student_id": {
            "title": "Student Id",
            "type": "string"
          }
        },
        "required": [
          "student_id",
          "name",
          "course_limit"
        ],
        "title": "StudentRegisteredWebhook",
        "type": "object"
      },
      "SubscribeStudentRequest": {
        "description": "Request body for the Subscribe Student command.",
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "string"
          }
        },
        "required": [
          "course_id"
        ],
        "title": "SubscribeStudentRequest",
        "type": "object"
      },
      "UnsubscribeStudentRequest": {
        "description": "Request body for the Unsubscribe Student command.",
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "string"
          }
        },
        "required": [
          "course_id"
        ],
        "title": "UnsubscribeStudentRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "A fake OAuth2 token: `<id>|<type>|<scopes>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/course-catalogue": {
      "get": {
        "description": "Return every registered course with its capacity and subscription count.",
        "operationId": "course_catalogue",
        "parameters": [
          {
            "description": "Answer only once the view has reached at least this position.",
            "in": "header",
            "name": "X-Position-AtLeast",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Answer only once the view has reached at least this position.",
              "title": "X-Position-Atleast"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogueResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Current-Position": {
                "description": "Position of the last event this view reflects. Absent when the view has processed nothing.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "425": {
            "description": "The view has not reached X-Position-AtLeast.",
            "headers": {
              "X-Current-Position": {
                "description": "Position of the last event this view reflects. Absent when the view has processed nothing.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Course Catalogue",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/register": {
      "post": {
        "description": "Register a new course with a capacity limit on concurrent subscriptions.",
        "operationId": "register_course",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "The command recorded nothing."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{course_id}/change-capacity": {
      "post": {
        "description": "Change the capacity limit on concurrent subscriptions for a course.",
        "operationId": "change_course_capacity",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeCourseCapacityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "The command recorded nothing."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Course Capacity",
        "tags": [
          "courses"
        ]
      }
    },
    "/livez": {
      "get": {
        "description": "Report whether this process can still recover on its own.",
        "operationId": "livez",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "A probed dependency is unavailable; `detail` names each one."
          }
        },
        "summary": "Livez",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/readyz": {
      "get": {
        "description": "Report whether this instance can serve correct traffic right now.",
        "operationId": "readyz",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "A probed dependency is unavailable; `detail` names each one."
          }
        },
        "summary": "Readyz",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/students/{student_id}/course-subscriptions": {
      "get": {
        "description": "Return a student's current course subscriptions.",
        "operationId": "student_course_subscriptions",
        "parameters": [
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "title": "Student Id",
              "type": "string"
            }
          },
          {
            "description": "Answer only once the view has reached at least this position.",
            "in": "header",
            "name": "X-Position-AtLeast",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Answer only once the view has reached at least this position.",
              "title": "X-Position-Atleast"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentCourseSubscriptionsResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Current-Position": {
                "description": "Position of the last event this view reflects. Absent when the view has processed nothing.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "425": {
            "description": "The view has not reached X-Position-AtLeast.",
            "headers": {
              "X-Current-Position": {
                "description": "Position of the last event this view reflects. Absent when the view has processed nothing.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Student Course Subscriptions",
        "tags": [
          "students"
        ]
      }
    },
    "/students/{student_id}/subscribe-to-course": {
      "post": {
        "description": "Subscribe a student to a course, subject to capacity and student limits.",
        "operationId": "subscribe_student",
        "parameters": [
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeStudentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "The command recorded nothing."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Subscribe Student",
        "tags": [
          "students"
        ]
      }
    },
    "/students/{student_id}/unsubscribe-from-course": {
      "post": {
        "description": "Unsubscribe a student from a course they are currently subscribed to.",
        "operationId": "unsubscribe_student",
        "parameters": [
          {
            "in": "path",
            "name": "student_id",
            "required": true,
            "schema": {
              "title": "Student Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsubscribeStudentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "The command recorded nothing."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unsubscribe Student",
        "tags": [
          "students"
        ]
      }
    },
    "/webhooks/student-registered": {
      "post": {
        "description": "Record a student registration reported by the registrar.",
        "operationId": "external_register_student",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRegisteredWebhook"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "The command recorded nothing."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "External Register Student",
        "tags": [
          "webhooks"
        ]
      }
    }
  }
}